FPGA connections, DAC linear regulator removed, DAC I2S breakout
Connected the SPI, I2S and GPIO to some (mostly) arbitrary pins. Notably
the SPI clock has to be connected to a clock pin on the FPGA.

The linear regulator is (hopefully) not needed for the DAC as we already
have a separate analog +3.3V created by the switching regulators.

The I2S wires to the DAC are now connected to the FPGA as well as
breakout pins over 0 ohm resistors, in case we need to connect to an
external FPGA.